fixed MSVC compile warning

+++ tutorials/viewer/viewer_device.cpp
@@ -264,7 +264,7 @@ RTCScene convertScene(ISPCScene* scene_in)
    /* use geometry instancing feature */
   if (g_instancing_mode == 1)
   {
-    for (size_t i=0; i<scene_in->numGeometries; i++)
+    for (unsigned int i=0; i<scene_in->numGeometries; i++)
     {
       ISPCGeometry* geometry = scene_in->geometries[i];
       if (geometry->type == SUBDIV_MESH) {
@@ -309,7 +309,7 @@ RTCScene convertScene(ISPCScene* scene_in)
   /* use scene instancing feature */
   else if (g_instancing_mode == 2 || g_instancing_mode == 3)
   {
-    for (size_t i=0; i<scene_in->numGeometries; i++)
+    for (unsigned int i=0; i<scene_in->numGeometries; i++)
     {
       ISPCGeometry* geometry = scene_in->geometries[i];
       if (geometry->type == SUBDIV_MESH) {
@@ -484,7 +484,7 @@ Vec3fa renderPixelStandard(float x, float y, const ISPCCamera& camera)
 {
   /* initialize sampler */
   RandomSampler sampler;
-  RandomSampler_init(sampler, x, y, 0);
+  RandomSampler_init(sampler, (int)x, (int)y, 0);
 
   /* initialize ray */
   RTCRay ray;
